Add test assumptions
This adds the ability to make assumptions about the state of the system, which if
incorrect, will result in an "Inconclusive" test outcome.

By assuming conditions rather than asserting on them, it is easier to find the
root cause of issues - especially  in the case where multiple tests are relying
on a common set of functions.

A new example project demonstrates how assumptions help to identify issues,
rather than all three example test functions failing, there will be a single
failed tests and two inconclusive tests.